pyproject.toml
[project]
name = "langgraph-agent-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-openai>=0.3.32",
    "langgraph>=0.6.6",
    "python-dotenv>=1.1.1",
    "requests>=2.31.0",
    "pandas>=2.0.0",
    "openrouter>=0.0.1",
]
